+++ src/main/java/com/marginallyclever/ro3/apps/RO3Frame.java
@@ -16,8 +16,8 @@ import com.marginallyclever.ro3.apps.logpanel.LogPanel;
 import com.marginallyclever.ro3.apps.nodedetailview.NodeDetailView;
 import com.marginallyclever.ro3.apps.nodetreeview.NodeTreeView;
 import com.marginallyclever.ro3.apps.webcampanel.WebCamPanel;
-import com.marginallyclever.ro3.render.OpenGLPanel;
-import com.marginallyclever.ro3.render.Viewport;
+import com.marginallyclever.ro3.apps.render.OpenGLPanel;
+import com.marginallyclever.ro3.apps.render.Viewport;
 import com.marginallyclever.robotoverlord.RobotOverlord;
 import com.marginallyclever.robotoverlord.swing.actions.AboutAction;
 import org.slf4j.Logger;
