+++ src/main/java/com/marginallyclever/ro3/apps/render/renderpasses/DrawCameras.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.ro3.render.renderpasses;
+package com.marginallyclever.ro3.apps.render.renderpasses;
 
 import com.jogamp.opengl.GL3;
 import com.jogamp.opengl.GLAutoDrawable;
@@ -6,11 +6,7 @@ import com.jogamp.opengl.GLContext;
 import com.marginallyclever.convenience.helpers.MatrixHelper;
 import com.marginallyclever.convenience.helpers.ResourceHelper;
 import com.marginallyclever.ro3.Registry;
-import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.nodes.Camera;
-import com.marginallyclever.ro3.node.nodes.HingeJoint;
-import com.marginallyclever.ro3.node.nodes.Pose;
-import com.marginallyclever.ro3.render.RenderPass;
 import com.marginallyclever.robotoverlord.systems.render.ShaderProgram;
 import com.marginallyclever.robotoverlord.systems.render.mesh.Mesh;
 import org.slf4j.Logger;
@@ -18,33 +14,36 @@ import org.slf4j.LoggerFactory;
 
 import javax.vecmath.Matrix4d;
 import javax.vecmath.Vector3d;
-import java.util.ArrayList;
-import java.util.List;
 
-public class DrawHingeJoints extends AbstractRenderPass {
-    private static final Logger logger = LoggerFactory.getLogger(DrawHingeJoints.class);
+/**
+ * Draws each {@link Camera} as a pyramid approximating the perspective view frustum.
+ */
+public class DrawCameras extends AbstractRenderPass {
+    private static final Logger logger = LoggerFactory.getLogger(DrawCameras.class);
     private final Mesh mesh = new Mesh();
-    private final Mesh circleFan = new Mesh();
     private ShaderProgram shader;
-    private int canvasWidth,canvasHeight;
-    private final float ringScale = 3;
 
-    public DrawHingeJoints() {
-        super("Hinge Joints");
+    public DrawCameras() {
+        super("Cameras");
 
+        // add mesh to a list that can be unloaded and reloaded as needed.
         mesh.setRenderStyle(GL3.GL_LINES);
-        mesh.addColor(1.0f,1.0f,1.0f,1);  mesh.addVertex(0,0,0);  // origin
-        mesh.addColor(1.0f,1.0f,1.0f,1);  mesh.addVertex(0,0,0);  // angle unit line
-
-        circleFan.setRenderStyle(GL3.GL_TRIANGLE_FAN);
-        circleFan.addColor(1.0f,1.0f,0.0f,0.25f);
-        circleFan.addVertex(0,0,0);  // origin
-        for(int i=0;i<=360;++i) {
-            float x = (float)Math.cos(Math.toRadians(i)) * ringScale;
-            float y = (float)Math.sin(Math.toRadians(i)) * ringScale;
-            circleFan.addColor(1.0f,1.0f,0.0f,0.25f);
-            circleFan.addVertex(x,y,0);
-        }
+        Vector3d a = new Vector3d(-1,-1,-1);
+        Vector3d b = new Vector3d( 1,-1,-1);
+        Vector3d c = new Vector3d( 1, 1,-1);
+        Vector3d d = new Vector3d(-1, 1,-1);
+        mesh.addColor(0,0,0,1);        mesh.addVertex(0,0,0);        mesh.addColor(0,0,0,1);        mesh.addVertex((float)a.x, (float)a.y, (float)a.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex(0,0,0);        mesh.addColor(0,0,0,1);        mesh.addVertex((float)b.x, (float)b.y, (float)b.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex(0,0,0);        mesh.addColor(0,0,0,1);        mesh.addVertex((float)c.x, (float)c.y, (float)c.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex(0,0,0);        mesh.addColor(0,0,0,1);        mesh.addVertex((float)d.x, (float)d.y, (float)d.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)a.x, (float)a.y, (float)a.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)b.x, (float)b.y, (float)b.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)b.x, (float)b.y, (float)b.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)c.x, (float)c.y, (float)c.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)c.x, (float)c.y, (float)c.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)d.x, (float)d.y, (float)d.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)d.x, (float)d.y, (float)d.z);
+        mesh.addColor(0,0,0,1);        mesh.addVertex((float)a.x, (float)a.y, (float)a.z);
     }
 
     @Override
@@ -63,7 +62,6 @@ public class DrawHingeJoints extends AbstractRenderPass {
     public void dispose(GLAutoDrawable glAutoDrawable) {
         GL3 gl3 = glAutoDrawable.getGL().getGL3();
         mesh.unload(gl3);
-        circleFan.unload(gl3);
         shader.delete(gl3);
     }
 
@@ -88,35 +86,15 @@ public class DrawHingeJoints extends AbstractRenderPass {
         shader.set1i(gl3,"diffuseTexture",0);
         gl3.glDisable(GL3.GL_DEPTH_TEST);
         gl3.glDisable(GL3.GL_TEXTURE_2D);
-        gl3.glDisable(GL3.GL_CULL_FACE);
 
-        List<Node> toScan = new ArrayList<>();
-        toScan.add(Registry.getScene());
-        while(!toScan.isEmpty()) {
-            Node node = toScan.remove(0);
-            toScan.addAll(node.getChildren());
-
-            if(node instanceof HingeJoint joint) {
-                double angle = joint.getAngle()-joint.getMinAngle();
-                float x = (float)Math.cos(Math.toRadians(angle)) * ringScale;
-                float y = (float)Math.sin(Math.toRadians(angle)) * ringScale;
-                mesh.setVertex(1,x,y,0);
-                mesh.updateVertexBuffers(gl3);
-
-                Pose pose = joint.findParent(Pose.class);
-                Matrix4d w = (pose==null) ? MatrixHelper.createIdentityMatrix4() : pose.getWorld();
-                Matrix4d rZ = new Matrix4d();
-                rZ.rotZ(Math.toRadians(joint.getMinAngle()));
-                w.mul(rZ);
-                w.transpose();
-                shader.setMatrix4d(gl3,"modelMatrix",w);
-                mesh.render(gl3);
-                int range = Math.max(0, (int)(joint.getMaxAngle()-joint.getMinAngle()) );
-                circleFan.render(gl3,1+range,0);
-            }
+        for(Camera cam : Registry.cameras.getList() ) {
+            // set modelView to world
+            Matrix4d w = cam.getWorld();
+            w.transpose();
+            shader.setMatrix4d(gl3,"modelMatrix",w);
+            mesh.render(gl3);
         }
 
         gl3.glEnable(GL3.GL_DEPTH_TEST);
-        gl3.glEnable(GL3.GL_CULL_FACE);
     }
 }
