+++ src/main/java/com/marginallyclever/ro3/apps/render/renderpasses/DrawHingeJoints.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.ro3.render.renderpasses;
+package com.marginallyclever.ro3.apps.render.renderpasses;
 
 import com.jogamp.opengl.GL3;
 import com.jogamp.opengl.GLAutoDrawable;
@@ -8,8 +8,8 @@ import com.marginallyclever.convenience.helpers.ResourceHelper;
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.nodes.Camera;
+import com.marginallyclever.ro3.node.nodes.HingeJoint;
 import com.marginallyclever.ro3.node.nodes.Pose;
-import com.marginallyclever.ro3.render.RenderPass;
 import com.marginallyclever.robotoverlord.systems.render.ShaderProgram;
 import com.marginallyclever.robotoverlord.systems.render.mesh.Mesh;
 import org.slf4j.Logger;
@@ -20,25 +20,29 @@ import javax.vecmath.Vector3d;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * Draw the ground plane.
- */
-public class DrawGroundPlane extends AbstractRenderPass {
-    private static final Logger logger = LoggerFactory.getLogger(DrawGroundPlane.class);
-    private final Mesh mesh = MatrixHelper.createMesh(1.0);
+public class DrawHingeJoints extends AbstractRenderPass {
+    private static final Logger logger = LoggerFactory.getLogger(DrawHingeJoints.class);
+    private final Mesh mesh = new Mesh();
+    private final Mesh circleFan = new Mesh();
     private ShaderProgram shader;
+    private int canvasWidth,canvasHeight;
+    private final float ringScale = 3;
 
-    public DrawGroundPlane() {
-        super("Ground plane");
+    public DrawHingeJoints() {
+        super("Hinge Joints");
 
         mesh.setRenderStyle(GL3.GL_LINES);
-        int v = 1000;
-        int stepSize=100;
-        for(int s=-v;s<v;s+=stepSize) {
-            mesh.addVertex(s, -v, 0);
-            mesh.addVertex(s,  v, 0);
-            mesh.addVertex( -v,s, 0);
-            mesh.addVertex(  v,s, 0);
+        mesh.addColor(1.0f,1.0f,1.0f,1);  mesh.addVertex(0,0,0);  // origin
+        mesh.addColor(1.0f,1.0f,1.0f,1);  mesh.addVertex(0,0,0);  // angle unit line
+
+        circleFan.setRenderStyle(GL3.GL_TRIANGLE_FAN);
+        circleFan.addColor(1.0f,1.0f,0.0f,0.25f);
+        circleFan.addVertex(0,0,0);  // origin
+        for(int i=0;i<=360;++i) {
+            float x = (float)Math.cos(Math.toRadians(i)) * ringScale;
+            float y = (float)Math.sin(Math.toRadians(i)) * ringScale;
+            circleFan.addColor(1.0f,1.0f,0.0f,0.25f);
+            circleFan.addVertex(x,y,0);
         }
     }
 
@@ -58,6 +62,7 @@ public class DrawGroundPlane extends AbstractRenderPass {
     public void dispose(GLAutoDrawable glAutoDrawable) {
         GL3 gl3 = glAutoDrawable.getGL().getGL3();
         mesh.unload(gl3);
+        circleFan.unload(gl3);
         shader.delete(gl3);
     }
 
@@ -74,19 +79,43 @@ public class DrawGroundPlane extends AbstractRenderPass {
         shader.setVector3d(gl3,"cameraPos",cameraWorldPos);  // Camera position in world space
         shader.setVector3d(gl3,"lightPos",cameraWorldPos);  // Light position in world space
         shader.setVector3d(gl3,"lightColor",new Vector3d(1,1,1));  // Light color
-        shader.set4f(gl3,"objectColor",1,1,1,0.125f);
+        shader.set4f(gl3,"objectColor",1,1,1,1);
         shader.setVector3d(gl3,"specularColor",new Vector3d(0.5,0.5,0.5));
         shader.setVector3d(gl3,"ambientLightColor",new Vector3d(0.2,0.2,0.2));
-        shader.set1i(gl3,"useVertexColor",0);
+        shader.set1i(gl3,"useVertexColor",1);
         shader.set1i(gl3,"useLighting",0);
         shader.set1i(gl3,"diffuseTexture",0);
+        gl3.glDisable(GL3.GL_DEPTH_TEST);
         gl3.glDisable(GL3.GL_TEXTURE_2D);
+        gl3.glDisable(GL3.GL_CULL_FACE);
 
-        Matrix4d w = MatrixHelper.createIdentityMatrix4();
-        w.transpose();
-        shader.setMatrix4d(gl3,"modelMatrix",w);
-        mesh.render(gl3);
+        List<Node> toScan = new ArrayList<>();
+        toScan.add(Registry.getScene());
+        while(!toScan.isEmpty()) {
+            Node node = toScan.remove(0);
+            toScan.addAll(node.getChildren());
+
+            if(node instanceof HingeJoint joint) {
+                double angle = joint.getAngle()-joint.getMinAngle();
+                float x = (float)Math.cos(Math.toRadians(angle)) * ringScale;
+                float y = (float)Math.sin(Math.toRadians(angle)) * ringScale;
+                mesh.setVertex(1,x,y,0);
+                mesh.updateVertexBuffers(gl3);
+
+                Pose pose = joint.findParent(Pose.class);
+                Matrix4d w = (pose==null) ? MatrixHelper.createIdentityMatrix4() : pose.getWorld();
+                Matrix4d rZ = new Matrix4d();
+                rZ.rotZ(Math.toRadians(joint.getMinAngle()));
+                w.mul(rZ);
+                w.transpose();
+                shader.setMatrix4d(gl3,"modelMatrix",w);
+                mesh.render(gl3);
+                int range = Math.max(0, (int)(joint.getMaxAngle()-joint.getMinAngle()) );
+                circleFan.render(gl3,1+range,0);
+            }
+        }
 
         gl3.glEnable(GL3.GL_DEPTH_TEST);
+        gl3.glEnable(GL3.GL_CULL_FACE);
     }
 }
